+++ backend/canisters/community/impl/src/updates/send_message.rs
@@ -99,7 +99,7 @@ fn c2c_send_message_impl(args: C2CArgs, state: &mut RuntimeState) -> C2CResponse
     };
 
     // Bots can't call this c2c endpoint since it skips the validation
-    if matches!(caller, Caller::Bot(_) | Caller::BotV2(_)) {
+    if matches!(caller, Caller::Bot(_) | Caller::BotCommand(_)) {
         return NotAuthorized;
     }
 
@@ -256,7 +256,7 @@ fn process_send_message_result(
 
                 if let Some(channel) = state.data.channels.get(&channel_id) {
                     for user_id in users_mentioned.all_users_mentioned {
-                        if user_id != caller.initiator()
+                        if caller.initiator().map(|i| i != user_id).unwrap_or_default()
                             && channel.chat.members.get(&user_id).is_some_and(|m| !m.user_type().is_bot())
                         {
                             activity_events.push((user_id, MessageActivity::Mention));
@@ -275,7 +275,7 @@ fn process_send_message_result(
                             thread_root_message_index,
                             replying_to_event_index.into(),
                         ) {
-                            if message.sender != caller.initiator()
+                            if caller.initiator().map(|i| i != message.sender).unwrap_or_default()
                                 && channel
                                     .chat
                                     .members